Refactor functionallity into separate files.
